Create FrameHandlerFactory interface
Fixes #11
